move user and password model out of repository layer
